// src/types/generated/StreamMagicManagerStateMsg.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { StreamMagicDevice } from "./StreamMagicDevice";
import type { WebSocketClientStatus } from "./WebSocketClientStatus";

export interface StreamMagicManagerStateMsg { devices: Array<StreamMagicDevice>, is_activating: boolean, is_discovering: boolean, is_testing_connection: boolean, websocket_client_status: WebSocketClientStatus, }