// src/types/generated/StreamMagicDevice.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export interface StreamMagicDevice { friendly_name: string, model: string, model_number: string | null, serial_number: string | null, url: string, udn: string, is_activating: boolean, is_active: boolean, }