// src/types/generated/WebSocketClientStatus.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { WebSocketClientConnectedDetails } from "./WebSocketClientConnectedDetails";
import type { WebSocketClientDisconnectedDetails } from "./WebSocketClientDisconnectedDetails";

export type WebSocketClientStatus = { "state": "Disconnected", "metadata": WebSocketClientDisconnectedDetails } | { "state": "Connecting", "metadata": string } | { "state": "Connected", "metadata": WebSocketClientConnectedDetails } | { "state": "TestingConnection" };