// src/types/generated/streammagic_payloads/Presets.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PresetItem } from "./PresetItem";

export interface Presets { start: number | null, end: number | null, max_presets: number | null, presettable: boolean | null, presets: Array<PresetItem> | null, }