// src/types/generated/streammagic_payloads/ZoneNowPlaying.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ZoneNowPlayingDisplay } from "./ZoneNowPlayingDisplay";
import type { ZoneNowPlayingQueue } from "./ZoneNowPlayingQueue";
import type { ZoneNowPlayingSource } from "./ZoneNowPlayingSource";

export interface ZoneNowPlaying { state: string | null, source: ZoneNowPlayingSource | null, display: ZoneNowPlayingDisplay | null, queue: ZoneNowPlayingQueue | null, controls: Array<string> | null, }