// src/types/generated/streammagic_payloads/ZonePlayState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ZonePlayStateMetadata } from "./ZonePlayStateMetadata";
import type { ZonePlayStateState } from "./ZonePlayStateState";

export interface ZonePlayState { state: ZonePlayStateState | null, position: number | null, presettable: boolean | null, queue_index: number | null, queue_length: number | null, queue_id: number | null, mode_repeat: string | null, mode_shuffle: string | null, metadata: ZonePlayStateMetadata | null, }