// src/types/generated/streammagic_payloads/ZoneState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export interface ZoneState { source: string | null, power: boolean | null, pre_amp_mode: boolean | null, pre_amp_state: string | null, mute: boolean | null, volume_step: number | null, volume_percent: number | null, volume_db: bigint | null, cbus: string | null, }