// src/types/generated/AmplifierManagerStateMsg.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AmplifierDevice } from "./AmplifierDevice";

export interface AmplifierManagerStateMsg { is_discovering: boolean, is_handling_amplifier: boolean, managed_device: AmplifierDevice | null, }